do not load Sigil's DEHACKED patch when auto-loading
Otherwise, this would break *any* episode finishing demo for Doom 1 if
Sigil.wad is auto-loaded at start. We already have the episode names
and par times hard-coded for Episode 5. Instead of removing the call
to A_BossDeath() from the last death states of Cyberdemon, Spider
Mastermind and Baron of Hell, we already have modified CheckBossEnd()
to return `false` for `gameepisode == 5`.