prevent the fuzz effect animation from remaining static
This happens if the number of pixels to apply the fuzz effect to is an
integer multiple of FUZZTABLE. As such, there is not much we can do
about this. But this way we make sure that the animation moves by at
least one pixel and never stands entirely still.

Fixes #490, thanks @JNechaevsky for the suggestion, although I
initially rejected it.