add the status bar face to the Crispy HUD
This provides an "intermediate Crispy HUD" with the status bar face
being visible. In order to prevent the face to blend into the weapon
sprite, the multiplayer face widget background is drawn beneath it with
translucency, so the weapon isn't entirely covered in this area.

There might still be some space left for improvements and discussion,
espacially regarding the background for the face widget, but I somehow
find the green multiplayer background for Player 1 quite logical.

Quite surely, there won't be a variant with a translucent face.

Fixes #485, thanks again @JNechaevsky and @Zodomaniac for your input
and especially UndeadRyker from the Doomworld Forums for finally
pushing me to implement this.