Correct Math for Loremaster Hook Effect
This code change fixes a bug in which the Loremaster's hook, which is meant to throw the player into the air, has nearly no effect on the player's Z momentum. Because of this, the player is not thrown into the air, only moved slightly in the x and y directions- rendering the Loremaster nearly harmless.

The problem was that the "temp" variable was being assigned a fixed point value (which is equivalent to a very large integer), then being divided by the target's mass (a normal integer, small in size), yielding a very large, nonsense number. The momz calculation calls for the (fixed point) z distance between the target and source to be divided by the temp value. However, since the (small) z distance is being divided by the (very large) nonsense temp value, the resulting momz is virtually zero almost every time.

This fix converts the player's mass to fixed point in the calculation and replaces "/" operators with proper fixed point math functions, yielding the correct z momentum.

To test the difference, warp to MAP27 to face the Loremaster, or use the GRAPPLE.SEH patch on the SeHackEd Doomworld page, which also uses the hook effect. In the vanilla Strife DOS exe and this fork, the player should properly be thrown into the air, while in the master branch as of writing, the player is only moved in the x and y directions.