draw the actual status bar face widget background in the Crispy Face HUD
This means, we don't draw the translucent background area in the
non-translucent HUD anymore, which may be considered a drawback. On
the pro side of things, the background now always looks exactly like
its pendant from the status bar, so we don't have to care about the
current game being a netgame or not nor color translations anymore.