doom: correctly determine the par time for MAP33
The par time for MAP33 is read from beyond the cpars[] array and
corresponds to the first four bytes ("Gamm") of the GAMMALVL0 string
literal interpreted as a 32-bit integer. This has been checked and
proven for the three regular DOOM.EXEs (i.e. 1.9, Ultimate, Final).

Fixes #1209.

Thanks @turol for the pointer to use memcpy() instead of pointer
casting.

While at it, change the data type for leveltime and partime to type
short in the statdump code to match the output of the DOS executable.