fix crash when the flag for the berserk pack sprite patch memory zone is changed
The sprite patch for the Berserk pack is shown in the bottom left
corner of the Crispy HUD when the Berserk pack is enabled. For this,
the patch lump is loaded into zone memory with the PU_STATIC flag,
which means it is not purgeable.

However, when the Berserk pack is rendered in the game scene, it
sprite patch is loaded into zone memory with the PU_CACHE flag, which
means it is purgeable whenever further memory is needed. Since the
patch lump is alread loaded into zone memory, it is not loaded again,
but the same pointer is returned as before. However, the flag gets
changed, so the memory zone may get purged. In this case, the pointer
in the Crispy HUD code path will point to some random memory content
and cause the crash in V_DrawPatch().

Fixes #503, thanks @IsBebs for the bug report and @Zodomaniac,
@JNechaevsky and @turol for helping with the analysis!