improve weapon sprite bobbing calculation (again)
The weapon being in its firing state doesn't necessarily mean the
player being in attacking state, as @unRyker points out.

With this change we track the attack key/button being held down, i.e.
the *intention* to attack, which hopefully provides for a reasonable
visual feedback to the user. Interestingly, this is also how it was
implemented in the past, c.f.
https://github.com/fabiangreffrath/crispy-doom/commit/bf31f717623aee7d974a812d2e91b3065dd35d8b

If this doesn't help, I'll have to implement @kitchen-ace's idea and
track weapon states by their semantical meaning.

Opinions please! Fixes #571 for the time being.