auto-load masterlevels.wad as episode 3 (#593)
With this commit, masterlevels.wad will be auto-loaded alongside DOOM2.WAD just like NERVE.WAD was for the BFG Edition Doom2 IWAD. However, with this commit, this will work for *any* Doom 2 IWAD.

As before, the episode menu will be printed with the correct graphics lumps if only nerve.wad is auto-loaded and the IWAD is the one from the BFG Edition. Since there are no "The Master Levels" menu graphics lumps available for the episode menu in any IWAD, the menu will be printed with the HUD font once this WAD is auto-loaded. Also, since this change means we will end up with up to three copies of MAP01-MAP09 (even if no further PWAD is loaded), the logic has been changed to append 'M' to level lumps and level name graphics lumps from masterlevela.wad and 'N' to the same lumps from nerve.wad. Furthermore, some logic needed to be added to handle the case that only either of the two WADs id auto-loaded.

This change is massive! I hope it doesn't fall on my feet that hard. ;)

@Zodomaniac The next release will probably be 5.9.0 - but certainly not on this Friday. ;)