doom: Refactor chat macro defaults for const correctness
chat_macros is passed to M_BindStringVariable which might ovewrite
its contents so it can't be const. But the default values are constants
so it should be const. Solve this by moving defaults to their own list
and initializing chat_macros in D_BindVariables just before they're
bound to string variables.