Add pistol/wand start command line option (#659)
* Add pistolstart flag to common crispy variables

This flag is being added to facilitate automatic pistol (or wand)
starts.

* doom: Add -pistolstart command line option

When the user includes -pistolstart as a command line option, it causes
there to be an automatic "pistol start" when advancing from one level to
the next. Specifically, the player's health is reset to 100, their armor
to 0 and their inventory is reduced to the following: pistol, fists and
50 bullets. This option has no effect when recording a demo, playing
back a demo or when starting a network game.

* heretic: Add -wandstart command line option

Implement equivalent Doom "pistol start" functionality for Heretic.

* Simplify pistolstart logic

* Add pistol/wand start error condition

If the user uses -pistolstart or -wandstart when playing a demo,
recording a demo or playing a network game, call an error. If playing
back a demo from the demo reel, silently ignore pistolstart.

Note: In most cases this error won't occur until the first proper level
transition.

* Revert accidental whitespace change