fix string buffer size calculation in levelstat implementation
@kraflab In G_FormatLevelStatTime() str is a pointer to char, so its size
is the size of a pointer, i.e. 4 bytes on a 32 bit system or 8 bytes
on a 64 bit system, respectively. It is not the size of the char array it
points to! :wink:

This leads to time strings being cut off after the third char (the
fourth char being reserved for '\0') such as "0:2" instead of
"0:25.17". Thanks Eric Claus for reporting this.