Workaround SDL_WINDOW_FULLSCREEN_DESKTOP issues.
On some environments using SDL_WINDOW_FULLSCREEN_DESKTOP will
result in the program losing its fullscreen property when
switching workspaces, moving the program to a different
workspace and maybe some other scenarios.

This can be worked in the user's environment with:

  SDL_VIDEO_MINIMIZE_ON_FOCUS_LOSS=0

But its better to set the hint in chocolate-doom directly.