do not remove crushed non-bleeding monsters anymore
Formerly, we would remove crushed non-bleeding monsters (such as the
Lost Soul if the Colored Blood feature is enabled) off the map by
setting their state to S_NULL, which appears not to be as harmless as
it seemed back then. Now, we do as Vanilla does and set the crushed
monster's state to S_GIBS, but replace its sprite with the invisible
SPR_TNT1.

Thanks ZeroMaster010 for the bug report and sorry for the desyncing
demo!