composite all textures, whether single-patched or not  (#698)
* composite all textures, whether single-patched or not

With this commit, Crispy now composes all textures, whether they are
single-patched or not. It generates to composites per texture, one for
rendering mid-textures on 2S wall which may contain translucent areas,
and another one for rendering textures on 1S walls and the sky,
respectively.

No column is ever read directly from a patch lump anymore. This keeps
the Medusa bug fixed but also fixes graphical glitches with e.g.
single-patched textures that are shorter than the patch, etc.

* comments and clean-ups

* more clean-ups

* do not remove crushed non-bleeding monsters anymore

Formerly, we would remove crushed non-bleeding monsters (such as the
Lost Soul if the Colored Blood feature is enabled) off the map by
setting their state to S_NULL, which appears not to be as harmless as
it seemed back then. Now, we do as Vanilla does and set the crushed
monster's state to S_GIBS, but replace its sprite with the invisible
SPR_TNT1.

Thanks ZeroMaster010 for the bug report and sorry for the desyncing
demo!

* initialize opaque texture column offset earlier

* compose tall patch compliant texture columns

Once we pass the 254 pixel boundary, the topdelta value becomes relative.

* fix previous implementation of tall patch compliant texture column composition