build: use python3 exclusively
Python 2 is deprecated and is going to be removed from Linux
distributions in the short term [1,2]. Thus, (parts of) the build
process will fail on current systems like e.g. Debian unstable.

The /bin/python binary is not going to exist anymore. The canonical
name for the Python 3 binary is, you guess it, /bin/python3. So, let's
just use this during our build process and go future proof.

Fixes #1158.

[1] https://wiki.debian.org/Python#Python_in_Debian
[2] https://wiki.debian.org/Python/2Removal