round SCREENWIDTH *down* to the nearest multiple of 4 (in hires mode)
The widescreen assets are 426 px wide. Divided by an (aspect ratio
corrected) screenheight of 240 px, this results in a ratio of 1.775
which is only slightly smaller than 16:9 = 1.7777.

In hires mode, for a 16:9 ratio, we want screenwidth to result in
(426 * 2 =) 852 px: 16 * 480 / 9 = 853.3333, which when rounded *down*
to the nearest multiple of 4 is 852. The same results are achieved
with common screen resolutions such as 1920x1080 or 2560x1440.

In lores mode, we cannot get away with a screenwidth of 426 px, because
this is not a multiple of 4. Thus, we still round screenwidth *up* to
428 px and end up with a 1 px pillar to the left and right.

Fixes #704, thanks @buvk.