Fix incorrect use of PRIiPTR/PRIuPTR
This is needed to support CHERI, and thus Arm's experimental Morello
prototype, where pointers are implemented using unforgeable capabilities
that include bounds and permissions metadata to provide fine-grained
spatial and referential memory safety, as well as revocation by sweeping
memory to provide heap temporal memory safety.

PRIiPTR and PRIuPTR and are for (u)intptr_t arguments which, for CHERI,
are implemented as capabilities in order to preserve pointer provenance
(converting a capability to an integer is a lossy operation). However,
these arguments are of type size_t or ptrdiff_t, which remain plain
integers. Thus, use the correct C99 format strings for these.