multiplayer: Fix integer/pointer confusion
CHERI, and thus Arm's experimental Morello prototype, implements
pointers using unforgeable capabilities that include bounds and
permissions metadata to provide fine-grained spatial and referential
memory safety, as well as revocation by sweeping memory to provide heap
temporal memory safety.

The existing code did in fact work for CHERI, but was inefficient, using
a full 128-bit capability (or 64-bit on systems with a 32-bit address
space) for values that are in fact just plain ints. The only need for
intptr_t is when casting to void * for TXT_SignalConnect's user_data
argument in order to silence warnings.