hexen: Add -gameversion, accepting 1.1 and 1.1r2 as possible values.
This is mostly based on code previously added for Doom. The main change
is that the behaviors of two variations of Hexen v1.1 are now covered.
They differ by the presence of a bugfix in A_SoAExplode, observed in DK
with -nomonsters. There is also an adjustment of HEXEN_VERSIONTEXT.

These changes are otherwise separate from support for the demo
version's IWAD, for which no separate game version was added.
As with Doom, the emulated version, 'Hexen 1.1' or
'Hexen 1.1 (alt)', is now printed on startup.