Hardcoding insane limits for heretic
So I have decided to hardcode limits that exceed DOOM32 and remove the option to toggle vanilla limits.  I will do this for the rest of the games in Sprinkled Doom.  The whole point of this port is to have large limits but stay 'Choco'.  So if people don't use the increased limits, whats the point?  These limits far exceed DOOM32 as I am multiplying them by the factor.