Heretic: brightmaps (#878)
* Initial part, wall/plane brightmaps

* STbar weapons

No Hell Staff at the moment. It shares a lot of bright pixels on red energy and  on staff itself, which makes bmap looking ugly. Without that bright pixels there will be notable "holes" on red energy... Probably better just leave it as is.

Same problem with one bright pixel on  Ethereal Crossbow, but at least it is possible to avoid it.

* Remove "doom" reference

* Artifacts and world weapons

* Light sources

No wall torch, it shares same colors in flame and lower metal part.

Serpent Torch have a face lit, but it looks okay (more or less), should be discussed.

* Monsters and Volcano

Excluded monster: Disciple of D'Sparil, because last death frame (pile of purple energy) is also lit, which is looking odd.

* Switch mobj brightmaps to per-state bias

- Better bmap for Iron Lich.
- Mostly solved Hell Staff attacks. This is as far as I can go.

* Update r_thing as well

* Add menu item!

* Make brightmaps const

As recommended by @turol

* Add Heretic's `crispy_brightmaps` variable to Setup

* More brightmapped objects

- Flame and Infernal orbs (no need for half-bright, fire is okay with full brightness).
- Weredragon fireball
- Iron Lich flame wall
- Volcano fireball

* Remove TODOs and redundant line breaks