HH: Capture mouselook when recording demos (#898)
When recording demos, mouselook movements are now mapped to the "look"
variable also used for keyboard look. Compared to normal gameplay, the
mouselook fidelity is reduced. This is similar to how turning resolution
is lower when recording a vanilla demo.  Furthermore, the amount of
travel in a tic is limited to what could be achieved using the keyboard.