Introduce Windows console wrapper
Add a standalone wrapper which is to be compiled as a non-gui
application and then renamed to chocolate-<game>.com. When
"chocolate-<game>" (with no file extension) is called from a Windows
terminal environment, Windows will run the .com file instead of the
.exe. This .com file will then call the actual chocolate-<game>.exe in a
child process.

The .com file sets an environment variable which can then be checked in
the main executable to determine if it was launched from the command
line. If so, attach to the console and redirect output to the calling
terminal.