Doom: various multiplayer fixes (#921)
* Fix incorrect demo progress bar length

Nothing especial, straight from Woof. Note that numplayersingame is still there, I will need for next fixes.

* Fix incorrect demo timer speed and value

Again, if there are more than one player.

Without this fix, the more players in the game, the faster timer is. Not sure if it's optimal approach, but it's working for both single and multi player demos.

* Interpolate other player arrows

Fixes notable jittering of other player arrows.

* Interpolate and fix other player arrows angle

* Fixes incorrect arrow angle in automap rotate mode. In static mode it's fine.

Also reduced scope of "smoothangle" in AM_drawPlayers.
It's not needed to be available for whole function.

* Show status bar values/facebg of choosen player in spy mode

Notes:
- Sound emmiting of displayplayer, not consoleplayer.
- Displayed messages of displayplayer, not consoleplayer.
- Use Lee's code for player face background.

* Keep automap open after dying in demoplaying

Not a fix of bug, really. Keeping automap open is handy for viewing how player arrows are messing all around!

* Fix indentation

* Make numplayersingame non global

Also replaced consoleplayer with displayplayer in HU_Start (wasn't commited earlier?).