strife: Add misc. Crispy features (#1019)
* strife: Support PWAD autoload directories

* strife: Disable "-merge" warning

* strife: Add clean screenshot key

* strife: Add misc. sound fixes

* strife: Add "Play Sounds in Full Length" option

* strife: Preserve targets/tracers in saved games

* strife: Preserve monster skill when loading a game

* strife: Fix popup screen resizing

When showing a popup, decrease the screen size to show the status bar only if the original fullscreen HUD is being used. Then, when the popup closes, increase the screen size to show the original fullscreen HUD again. This preserves the Crispy HUD view and any screen size changes made while the popup was active.