Hexen: add true color support  (#1145)
* Initials

* Vissprite blending function

* Minimal support for foggy maps and gamma-correction

* Implement ice and Wraithverge palettes

* Implement Bloodscourge palettes

Also fine-tune Wraithverge a little bit.

* Implement poison palettes

This is exactly what happens when the vanilla bug with the missing first palette plays on our side, because index 13 uses a different colouring!

* Fine-tune color panes for better representation of paletted render

Still not 100% identical to vanilla, but very close. Probably it will not be possible to see the difference without comparing frame by frame.

* Fine-tune colors even more using DoomWiki values

Good to know I was very close to them! 👩🏻‍🎨

Co-Authored-By: kitchen-ace <47063032+kitchen-ace@users.noreply.github.com>

* Refactor foggy colormaps handling

First of all, reduce diff and get rid of "actual_colormap" external variable as we always know exactly which colormap should be used.

Next, since things are gettig more complicated in true color, it might be a good idea to do not alter original R_InitColormaps function and use our own instead. Original one doesn't need any incoming parameters after all.

Colour translations and string tables are now generated only once at startup, we don't need to regenerate them every level change and every gamma-correction change (worth to follow this suite for other games in other PR).

Finally, foggy maps now have decent fog, but this seems to be the last nail in the coffin of "support for modified colormaps"...

* Do not forget to release PLAYPAL lump name

It's locked at the start of R_InitTrueColormaps and must be released in the end, since commonly it is loaded into PU_CACHE!

* implement "Alt" tinttab blending

* Review and slightly simplify Raven patch drawing functions

In this commit:

- V_DrawShadowedPatch: shadows now uses proper values for paletted render and more accurate opacity for true colour (that's why they were not intense enough).
- V_DrawAltTLPatch: now uses proper translucency levels for paletted and true colour renders (drawalttinttab).
- There is no need of arrays for Raven translucent patches, as they will never be coloured (dp_translation) or have different translucencies (dp_translucent), let's just use straight and simple declarations.

* Implement true color fading for finale screens

* Fix accidental delete

* Handle vanilla bug with incorrect 255th COLORMAP color

And provide reasonable support for custom COLORMAP lump: do not replace black with white if custom COLORMAP is loaded.

Co-Authored-By: Fabian Greffrath <fabian@greffrath.com>

* Translucensy for Cleric's Icon of the Defender effects

This is where difference between common and "alt" blending functions is very notable.

* bring (non-)truecolor colormaps in line with Vanilla Hexen

* truecolor Hexen: draw patches directly with palette colors

* Remove no longer used code, macrocize BLENDSTEP

* Guard alt tinttab sprite blending with "else"

---------

Co-authored-by: kitchen-ace <47063032+kitchen-ace@users.noreply.github.com>
Co-authored-by: Fabian Greffrath <fabian@greffrath.com>